backend/requirements.txt
fastapi
uvicorn
joblib
python-dotenv
scikit-learn
pydantic
pytest
